backend/requirements.txt
langchain==0.3.7
langchain-core==0.3.17
langchain-community==0.3.7

pinecone==4.1.1
langchain-huggingface==0.1.0
sentence-transformers==3.1.1

langchain-google-genai==2.0.6
google-generativeai==0.8.3

langgraph==0.2.36

pydantic==2.9.2
pydantic-settings==2.4.0

python-dotenv==1.0.1

fastapi==0.115.0
uvicorn==0.30.6

tiktoken==0.7.0
protobuf==4.25.3

unstructured==0.14.9
pypdf==5.1.0

